presentation-api/controlling-ua/PresentationRequest_onconnectionavailable-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Presentation.onconnectionavailable (manual test)</title>
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs">
<link rel="help" href="https://w3c.github.io/presentation-api/#starting-a-presentation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p>Click the button below and select the available presentation display, to start the manual test.</p>
<button id="presentBtn" onclick="startPresentation()">Start Presentation Test</button>


<script>
    // ----------
    // DOM Object
    // ----------
    var presentBtn = document.getElementById("presentBtn");

    // ------------
    // Request init
    // ------------
    var validUnixDate = new Date().getTime() + String(Math.floor(Math.random() * 1e5)),
            validPresURL = '../receiving-ua/idlharness.html#__castAppId__=2334D33A/__castClientId__=' + validUnixDate,
            request = new PresentationRequest(validPresURL);

    // --------------------------------------------------------------------------
    // Start New PresentationRequest.onconnectionavailable Test (success) - begin
    // --------------------------------------------------------------------------
    var startPresentation = function () {
        async_test(function (t) {
            request.onconnectionavailable = t.step_func_done(function (evt) {
                var connection = evt.connection;

                // Check the initial state of a presentation connection
                test(function() {
                    assert_equals(connection.state, 'connecting');
                }, 'The initial state of the presentation connection is "connecting".');

                // Check, if the connection ID is set
                test(function () {
                    assert_true(!!connection.id);
                }, 'The connection ID is set.');

                // Check the type of the connection.id
                test(function () {
                    assert_true(typeof connection.id === 'string');
                }, 'The connection ID is a string.');

                // Check the instance of the connection
                test(function () {
                    assert_true(connection instanceof PresentationConnection);
                }, 'The connection is an instance of PresentationConnection.');
            });

            request.start();
        }, 'The presentation was started successfully.');
    }
    // ------------------------------------------------------------------------
    // Start New PresentationRequest.onconnectionavailable Test (success) - end
    // ------------------------------------------------------------------------
</script>
